Updated remove bday functionality
1. game stats can now be called after a remove action has been ran
2. renamed game stats function to getUpcomingBirthdays
3. moved some logic out of events into ui so that game stats could be called after remove
